vector::size_type is 64bits on x64, dont pass it to printf("%d")
git-svn-id: https://svn.isis.vanderbilt.edu/MoBIES/UDM/trunk@4082 c0adbae9-c137-0410-b742-ec4b2d2ceded

+++ src/UdmOclPat/GOCL/src/OCLObjectExBasic.cpp
@@ -21,7 +21,7 @@ namespace OclBasic
 	std::string PrintCollection( const std::string& strTypeName, const OclMeta::ObjectVector& vecElements )
 	{
 		char chBuffer[ 100 ];
-		sprintf( chBuffer, "%d", vecElements.size() );
+		sprintf( chBuffer, "%d", (int)vecElements.size() );
 		return strTypeName + " { size: " + std::string( chBuffer ) + "; } ";
 		/*
 		std::string strObject = strTypeName + " { " + std::string( chBuffer ) + " , [ ";
